Added search function
Added new search function. This function takes a value via user input and alerts them if that value is currently in the stack. If the value is on the stack, the user is also told at which position in the stack that value is, with position 0 corresponding to the top of the stack.